\name{Add7Multiply-package}
\alias{Add7Multiply-package}
\alias{Add7Multiply-package}
\docType{package}
\title{Add 7 and multiply}
\description{The package allows to use Add7Multiply function. See the documentation on the function Add7Multiply.}
\details{
\tabular{ll}{
Package: \tab Add7Multiply\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-12-10\cr
License: \tab GPL >= 2\cr
}
}
\author{
Yulia Kononchuk
Maintainer: Yulia Kononchuk <kononchuk@rsm.nl>
}
\references{
}
\keyword{package}
\seealso{
}
\examples{
v1 <- c(3,4,5)
v2 <- c(10,20)
result <- Add7Multiply (v1, v2) 
print(result)}
